@@ -4,19 +4,16 @@
   "description": "",
   "license": "",
   "sideEffects": false,
+  "type": "module",
   "scripts": {
-    "dev:css": "tailwindcss -w -i ./styles/app.css -o app/gen-styles/app.css",
-    "dev:remix": "remix watch",
-    "dev:wrangler": "cross-env NODE_ENV=development npx wrangler pages dev ./public --binding $(cat .env) --kv KV --port 3000",
-    "dev": "remix build && run-p dev:*",
-    "start": "cross-env NODE_ENV=production npm run dev:wrangler",
-    "build:css": "tailwindcss -m -i ./styles/app.css -o app/gen-styles/app.css",
-    "build": "npm run build:css && remix build"
+    "build": "remix build",
+    "start": "pnpm wrangler pages dev --compatibility-date=2023-12-30 ./public --binding $(cat .env) --kv KV --port 3000",
+    "dev": "remix dev --manual -c \"pnpm start\""
   },
   "dependencies": {
-    "@remix-run/cloudflare": "^1.11.1",
-    "@remix-run/cloudflare-pages": "^1.11.1",
-    "@remix-run/react": "^1.11.1",
+    "@remix-run/cloudflare": "^2.4.1",
+    "@remix-run/cloudflare-pages": "^2.4.1",
+    "@remix-run/react": "^2.4.1",
     "cross-env": "^7.0.3",
     "date-fns": "^2.29.3",
     "isbot": "^3.6.5",
@@ -25,12 +22,12 @@
     "zod": "^3.19.1"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^3.14.1",
-    "@remix-run/dev": "^1.11.1",
-    "@remix-run/eslint-config": "^1.11.1",
-    "@tailwindcss/forms": "^0.5.3",
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
+    "@cloudflare/workers-types": "^4.20231218.0",
+    "@remix-run/dev": "^2.4.1",
+    "@remix-run/eslint-config": "^2.4.1",
+    "@tailwindcss/forms": "^0.5.7",
+    "@types/react": "^18.2.46",
+    "@types/react-dom": "^18.2.18",
     "autoprefixer": "^10.4.7",
     "concurrently": "^7.2.0",
     "eslint": "^8.15.0",
@@ -38,11 +35,11 @@
     "postcss": "^8.4.14",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
-    "tailwindcss": "^3.2.4",
-    "typescript": "^4.9.4",
-    "wrangler": "^2.8.1"
+    "tailwindcss": "^3.4.0",
+    "typescript": "^5.3.3",
+    "wrangler": "^3.22.1"
   },
   "engines": {
-    "node": ">=16.17.0"
+    "node": ">=18.0.0"
   }
 }
