@@ -1,16 +1,12 @@
-/**
- * @type {import('@remix-run/dev').AppConfig}
- */
-module.exports = {
-  serverBuildTarget: "cloudflare-pages",
-  server: "./server.js",
-  devServerBroadcastDelay: 1000,
+/** @type {import('@remix-run/dev').AppConfig} */
+export default {
   ignoredRouteFiles: ["**/.*"],
-	future: {
-    v2_routeConvention: true,
-  },
-  // appDirectory: "app",
-  // assetsBuildDirectory: "public/build",
-  // serverBuildPath: "functions/[[path]].js",
-  // publicPath: "/build/",
+  server: "./server.ts",
+  serverBuildPath: "functions/[[path]].js",
+  serverConditions: ["workerd", "worker", "browser"],
+  serverDependenciesToBundle: "all",
+  serverMainFields: ["browser", "module", "main"],
+  serverMinify: true,
+  serverModuleFormat: "esm",
+  serverPlatform: "neutral",
 };
