@@ -1,12 +1,11 @@
 import { Link } from '@remix-run/react';
 import { useState } from 'react';
-import { posts } from 'app/routes/blog._index';
 
-export const meta = () => ({
-	title: 'Dan Gurney - Personal Site',
-});
-
-const mostRecentPost = posts[0];
+export const meta = () => [
+	{
+		title: 'Dan Gurney - Personal Site',
+	},
+];
 
 export default function Index() {
 	const [showYears, setShowYears] = useState(false);
